+++ packages/fighting-design/page-header/src/page-header.ts
@@ -1,8 +1,8 @@
-import type { ExtractPropTypes, PropType, VNode } from 'vue'
+import type { ExtractPropTypes, PropType, VNode, Component } from 'vue'
 
 export const Props = {
   icon: {
-    type: Object as PropType<VNode>,
+    type: Object as PropType<VNode | Component>,
     default: (): null => null
   },
   iconSize: {