+++ packages/fighting-design/toolbar-item/src/toolbar-item.ts
@@ -1,4 +1,4 @@
-import type { PropType, ExtractPropTypes, VNode } from 'vue'
+import type { PropType, ExtractPropTypes, VNode, Component } from 'vue'
 
 export const Props = {
   color: {
@@ -6,7 +6,7 @@ export const Props = {
     default: (): string => ''
   },
   icon: {
-    type: Object as PropType<VNode>,
+    type: Object as PropType<VNode | Component>,
     default: (): null => null
   },
   iconSize: {